+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2025-11-14
+
+### Fixed
+
+- Fix unreleased section custom changelog sub section types
+
 ## [2.0.0] - 2025-11-09
 
 ### Fixed
 
-- Fixed newline changelog token -> 
+- Fix newline changelog token
 
 ### Added